Move out scaffold controls
